changed author back to poster
please note, don't change variable names in entities. These variable are
column names. If you change them, the database performs an alter table
query and tries to find a column name that doesn't exist. This is very
bad. I changed the name back to 'poster'. A poster is someone who posts
something on the site. The poster posts a post. The poster posts
comments. Comments and posts are posted by the poster.